+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## [Unreleased]
 ### Changed
 - Changed #tag# to ignore if the last char inside is space (like #tag #tag)
+- Removed unnecessary logging in tenshi provider